Fix shader error when building projects
