+++ source/com.unity.cluster-display.graphics/Editor/EnsureShadersIncluded.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04e00f2473e67aa49b900f568af4fe04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
